+++ lib/src/status.dart
@@ -1,0 +1,6 @@
+enum WStoreStatus {
+  init,
+  loading,
+  loaded,
+  error,
+}