+++ lib/wstore.dart
@@ -1,5 +1,6 @@
 library wstore;
 
+export 'src/status.dart';
 export 'src/builder.dart';
 export 'src/context.dart';
 export 'src/error.dart';